==> LCMS/IsolatesGerm/04_NMDS_IsolatesGerm.R <==
#NMDS
#Normalized matrix
#LCMS Isolates Germ

#Taylor Portman
#22NOV22

#Import libraries
library(readxl)
library(ggpubr)
library(ggsci)
library(gridExtra)
library(vegan)
library(factoextra)
library(vsn)
library(tidyverse)
source('/Volumes/GoogleDrive/My Drive/projects/MastersThesis/R/ERLE_fungal_symbionts/LCMS/functions/functions_cdis_norm_stats.R')

#Set directories
project_dir <- setwd("/Volumes/GoogleDrive/My Drive/projects/MastersThesis/R/ERLE_fungal_symbionts/LCMS/IsolatesGerm")
figures_dir <- file.path(project_dir, paste0('RP_figures'))
tables_dir <- file.path(project_dir,  paste0('RP_data_clean'))
# Import data -------------------------------------------------------------
norm.matrix<-read.csv('RP_data_clean/normalized_transformed_auc_table.csv', row.names = 1)
metadata<-read.csv("RP_data_clean/metadata.csv")

#Set distance matrix method for relative abundance -ra calculations
nmds.matrix <- norm.matrix
dm.method <- 'bray'
dm <- vegdist(nmds.matrix, method=dm.method)

# NMDS  -------------------------------------------------------------------
color<- c( "#E64B35B2", "#4DBBD5B2",  "#00A087B2",  '#7E6148B2')
set.seed(123)
nmds <- metaMDS(dm,
                k = 2,
                maxit = 999,
                trymax = 500,
                wascores = TRUE)
stressplot(nmds)
# Extract nmds scores for plotting
nmds.scores <- as.data.frame(scores(nmds))
nmds.scores <- rownames_to_column(nmds.scores, var = 'SampleID')
nmds.scores <- left_join(nmds.scores, metadata, by = 'SampleID')
nmds_plot <- plot_nmds(nmds.scores, Plant, Fungi, color) +
  labs(title = 'NMDS plot by relative abundance')
nmds_plot
figure_file <- file.path(figures_dir, 'nmds_relative_abundance.png')
ggsave(figure_file, nmds_plot, dpi = 300)


# Permanova ---------------------------------------------------------------

rownames(metadata) <- metadata$SampleID
set.seed(456)
permanova <- adonis2(dm ~ Fungi+ Plant, 
                     data= metadata, 
                     permutations=999, 
                     method="bray")
permanova #Marginally Significant Plant p=0.096

permanova <- adonis2(dm ~ Fungi, 
                    data= metadata, 
                    permutations=999, 
                    method="bray")
permanova #Not sig: p=0.341
